% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_methods.R
\name{bestglm_ic}
\alias{bestglm_ic}
\title{Variable selection using the bestglm package}
\usage{
bestglm_ic(data, family, IC = "AIC")
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{IC}{information criteria to use: "AIC", "BIC", "BICg", "BICq", "LOOCV", "CV".}
}
\description{
Variable selection using the bestglm package
}
