% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_methods.R
\name{exhaustive_RDBC}
\alias{exhaustive_RDBC}
\title{Exhaustive variable selection using the RDBC}
\usage{
exhaustive_RDBC(data, family, tcc)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{tcc}{the tuning constant c in Huber's psi-function.}
}
\description{
Exhaustive variable selection using the RDBC
}
