% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information_Criterion.R
\name{matrix_IC}
\alias{matrix_IC}
\title{Calculate the Information Criterion for a robust model over a matrix}
\usage{
matrix_IC(matrix, data, family)
}
\arguments{
\item{matrix}{a boolean matrix indicating which variables to include on each row.}

\item{data}{a data frame containing the variables in the model.}

\item{family}{a description of the error distribution and link function to be used in the model.}
}
\description{
Calculate the Information Criterion for a robust model over a matrix
}
