% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{model_check}
\alias{model_check}
\title{A function to check whether the selected model is a correct or wrong model.}
\usage{
model_check(model, true)
}
\arguments{
\item{model}{a vector containing which variables were selected.}

\item{true}{a vector containing all the variables in the true model.}
}
\description{
A function to check whether the selected model is a correct or wrong model.
}
