% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_check.R
\name{model_check_RobStab}
\alias{model_check_RobStab}
\title{A function to check whether the selected model is a correct or wrong model for the RobStab method.}
\usage{
model_check_RobStab(list, true)
}
\arguments{
\item{list}{a list containing vectors for each subtractive lack-of-fit choice and which variables were selected.}

\item{true}{a vector containing all the variables in the true model.}
}
\description{
A function to check whether the selected model is a correct or wrong model for the RobStab method.
}
