% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_space.R
\name{model_space}
\alias{model_space}
\title{Find the candidate model space after B bootstraps}
\usage{
model_space(
  data,
  B,
  m,
  nStrata = 8,
  family,
  k = 1,
  resid = "pearson",
  coef = TRUE,
  wald = TRUE,
  dev = TRUE,
  bootstraps = NA,
  center = TRUE,
  centerFun = median,
  scale = TRUE,
  scaleFun = mad,
  tcc
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{B}{The number of bootstrap re-samples to run.}

\item{m}{The m used for the m-out-of-n boostrap.}

\item{nStrata}{Number of strat used for the stratified re-sampling.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{k}{The minimum number of times a mondel must be identified through resampling to be included in the rediced candidate space.}

\item{resid}{the type of residuals to be used for the stratified re-sampling.}

\item{coef}{A TRUE/FALSE value to indicate whether to evaluate RobStab using regression coefficients.}

\item{wald}{A TRUE/FALSE value to indicate whether to evaluate RobStab using wald statisitcs.}

\item{dev}{A TRUE/FALSE value to indicate whether to evaluate RobStab using deviances.}

\item{bootstraps}{Recreact results using previously used resamples.}

\item{centerFun}{a function to compute an estimate of the center of a variable. Defualt is the median.}

\item{scale}{A TRUE/FALSE value to indicate whether to scale the design matrix.}

\item{tcc}{the tuning constant c in Huber's psi-function.}

\item{centre}{A TRUE/FALSE value to indicate whether to center the design matrix.}

\item{a}{function to compute an estimate of the scale of a variable. Default is the MAD.}
}
\description{
Find the candidate model space after B bootstraps
}
