% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{step_glmrob}
\alias{step_glmrob}
\title{Stepwise selection using Robust Quasi-deviance.}
\usage{
step_glmrob(data, family, anovaTest = "QDapprox", pVal = 0.05, tcc)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{anovaTest}{a character string specifying the test statistic to be used. (Partially) matching one of "Wald", "QD" or "QDapprox".}

\item{pVal}{a number identifying the p-value threshold for the stepwise selection.}

\item{tcc}{the tuning constant c in Huber's psi-function.}
}
\description{
Stepwise selection using Robust Quasi-deviance.
}
