% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{step_ic}
\alias{step_ic}
\title{Stepwise model selection using MLE based methods.}
\usage{
step_ic(data, family, k = 2)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{k}{the multiple of the number of degrees of freedom used for the penalty. Only k = 2 gives the genuine AIC: k = log(n) is sometimes referred to as BIC or SBC.}
}
\description{
Stepwise model selection using MLE based methods.
}
