% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_space.R
\name{model_space}
\alias{model_space}
\title{Find the candidate model space after B bootstraps}
\usage{
model_space(
  data,
  B,
  m,
  nStrata = 8,
  family,
  k = 1,
  resid = "pearson",
  coef = TRUE,
  wald = TRUE,
  dev = TRUE,
  bootstraps = NA,
  tcc
)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}
}
\description{
Find the candidate model space after B bootstraps
}
