% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_values.R
\name{s_values}
\alias{s_values}
\title{Compute the subtractive lack-of-fit measures}
\usage{
s_values(weights, data, n, p, family, coef, wald, dev)
}
\arguments{
\item{data}{a data frame containing the variables in the model.}
}
\description{
Compute the subtractive lack-of-fit measures
}
